JBIDE-15520 - Clean UI and Core dependencies
Also moved some UI tests in a new org.jboss.tools.ws.jaxrs.ui.test project
