+++ plugins/org.jboss.tools.ws.jaxrs.core/META-INF/MANIFEST.MF
@@ -14,11 +14,8 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.7.0";visibility:=reex
  org.eclipse.jdt.launching.macosx;bundle-version="3.2.100";resolution:=optional;visibility:=reexport,
  org.eclipse.jface.text;bundle-version="3.7.0";visibility:=reexport,
  org.eclipse.jst.j2ee.ejb;bundle-version="1.1.502",
- org.eclipse.ui.workbench;bundle-version="3.7.0";visibility:=reexport,
  org.eclipse.jdt.core.manipulation;bundle-version="1.3.0";visibility:=reexport,
- org.eclipse.ui.editors;bundle-version="3.7.0";visibility:=reexport,
  org.eclipse.core.commands;bundle-version="3.6.0";visibility:=reexport,
- org.eclipse.ui;bundle-version="3.7.0";visibility:=reexport,
  org.eclipse.ltk.core.refactoring;bundle-version="3.5.200";visibility:=reexport,
  org.eclipse.wst.validation;bundle-version="1.2.300";visibility:=reexport,
  org.eclipse.jst.j2ee.web;bundle-version="1.1.301";visibility:=reexport,
