+++ plugins/org.jboss.tools.ws.jaxrs.core/plugin.properties
@@ -2,7 +2,3 @@ PLUGIN_NAME=JBoss JAX-RS Tooling (Core)
 PLUGIN_PROVIDER=JBoss by Red Hat
 VALIDATOR_NAME=JAX-RS Validator
 
-
-ConfigureMenu_AddJaxrs11Support=Add JAX-RS 1.1 support...
-ConfigureMenu_RemoveJaxrs11Support=Remove JAX-RS 1.1 support...
-
