+++ plugins/org.jboss.tools.ws.jaxrs.ui/META-INF/MANIFEST.MF
@@ -43,3 +43,4 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.7.0",
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ClassPath: .
+Export-Package: org.jboss.tools.ws.jaxrs.ui.configuration;x-friends:="org.jboss.tools.ws.jaxrs.ui.test"
