+++ plugins/org.jboss.tools.ws.jaxrs.ui/plugin.properties
@@ -7,3 +7,7 @@ PreferencePage_JaxrsSettings=JAX-RS Validator
 PreferencePage_Validator=JAX-RS Validator
 PreferencePage=JAX-RS
 
+ConfigureMenu_AddJaxrs11Support=Add JAX-RS Support...
+ConfigureMenu_RemoveJaxrs11Support=Remove JAX-RS Support...
+
+
