+++ plugins/org.jboss.tools.ws.jaxrs.ui/src/org/jboss/tools/ws/jaxrs/ui/configuration/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author xcoulon
+ *
+ */
+package org.jboss.tools.ws.jaxrs.ui.configuration;
