+++ tests/org.jboss.tools.ws.jaxrs.core.test/build.properties
@@ -5,5 +5,9 @@ bin.includes = META-INF/,\
                plugin.properties,\
                lib/hamcrest-all-1.3.0RC2.jar,\
                lib/mockito-core-1.9.0-rc1.jar,\
-               lib/objenesis-1.2.jar
+               lib/objenesis-1.2.jar,\
+               .,\
+               projects/
 jre.compilation.profile = JavaSE-1.6
+
+
