+++ tests/org.jboss.tools.ws.jaxrs.core.test/src/org/jboss/tools/ws/jaxrs/core/junitrules/WorkbenchTasks.java
@@ -150,7 +150,7 @@ public class WorkbenchTasks {
 	private static IPath getProjectSourcePath(String projectName) {
 		IPath path = null;
 		if (System.getProperty("user.dir") != null) {
-			path = new Path(System.getProperty("user.dir")).append("projects").append(projectName).makeAbsolute();
+			path = new Path(System.getProperty("user.dir")).removeLastSegments(1).append("org.jboss.tools.ws.jaxrs.core.test").append("projects").append(projectName).makeAbsolute();
 		} else {
 			Assert.fail("The sample project was not found in the launcher workspace under name '" + projectName + "'");
