+++ tests/org.jboss.tools.ws.jaxrs.ui.test/META-INF/MANIFEST.MF
@@ -3,11 +3,11 @@ Bundle-ManifestVersion: 2
 Bundle-Localization: plugin
 Bundle-Name: %PLUGIN_NAME
 Bundle-Vendor: %PLUGIN_PROVIDER
-Bundle-SymbolicName: org.jboss.tools.ws.jaxrs.core.test;singleton:=true
+Bundle-SymbolicName: org.jboss.tools.ws.jaxrs.ui.test;singleton:=true
 Bundle-Version: 1.7.0.qualifier
-Bundle-Activator: org.jboss.tools.ws.jaxrs.core.JBossJaxrsCoreTestsPlugin
+Bundle-Activator: org.jboss.tools.ws.jaxrs.ui.JBossJaxrsUITestsPlugin
 Require-Bundle: org.eclipse.core.runtime;bundle-version="3.7.0",
- org.jboss.tools.ws.jaxrs.core;bundle-version="1.2.2",
+ org.jboss.tools.ws.jaxrs.ui,
  org.eclipse.jdt.launching.macosx;bundle-version="3.2.100";resolution:=optional;visibility:=reexport,
  org.eclipse.ui.ide;bundle-version="3.7.0",
  org.slf4j.api;bundle-version="1.6.1",
@@ -33,13 +33,11 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.7.0",
  org.eclipse.wst.validation;bundle-version="1.2.300",
  org.apache.commons.io;bundle-version="2.0.1",
  org.jboss.tools.common.validation;bundle-version="3.4.0",
- org.jboss.tools.common.core;bundle-version="3.5.2"
+ org.jboss.tools.common.core;bundle-version="3.5.2",
+ org.jboss.tools.ws.jaxrs.core,
+ org.jboss.tools.ws.jaxrs.core.test;bundle-version="1.7.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ActivationPolicy: lazy
 Eclipse-RegisterBuddy: org.apache.log4j
 Eclipse-BuddyPolicy: registered
-Bundle-ClassPath: .,
- lib/hamcrest-all-1.3.0RC2.jar,
- lib/mockito-core-1.9.0-rc1.jar,
- lib/objenesis-1.2.jar
-Export-Package: org.jboss.tools.ws.jaxrs.core.junitrules;x-friends:="org.jboss.tools.ws.jaxrs.ui.test"
+Bundle-ClassPath: .
