+++ tests/org.jboss.tools.ws.jaxrs.ui.test/build.properties
@@ -1,0 +1,5 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               plugin.properties,\
+               .
