+++ tests/org.jboss.tools.ws.jaxrs.ui.test/plugin.properties
@@ -1,0 +1,3 @@
+PLUGIN_NAME=JBoss JAX-RS Tooling (UI Tests)
+PLUGIN_PROVIDER=JBoss by Red Hat
+
