+++ tests/org.jboss.tools.ws.jaxrs.ui.test/pom.xml
@@ -1,0 +1,100 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>org.jboss.tools.ws</groupId>
+		<artifactId>tests</artifactId>
+		<version>1.7.0-SNAPSHOT</version>
+	</parent>
+	<groupId>org.jboss.tools.ws.tests</groupId>
+	<artifactId>org.jboss.tools.ws.jaxrs.ui.test</artifactId>
+	<packaging>eclipse-test-plugin</packaging>
+
+	<properties>
+		<coverage.filter>org.jboss.tools.ws.jaxrs.ui*</coverage.filter>
+		<emma.instrument.bundles>org.jboss.tools.ws.jaxrs.ui</emma.instrument.bundles>
+	</properties>
+
+	<build>
+	  <plugins>
+            <plugin>
+              <groupId>org.eclipse.tycho</groupId>
+              <artifactId>tycho-surefire-plugin</artifactId>
+              <configuration>
+                  <includes>
+                    <include>**/*TestCase.class</include>
+                  </includes>
+                  <excludes>
+                    <exclude>**/Abstract*.class</exclude>
+                  </excludes>
+                </configuration>
+              </plugin>	
+	  </plugins>
+	</build>
+	<profiles>
+		<profile>
+			<!-- Ignore optional deps to jdt.launching.macosx -->
+			<id>ignore-mac-deps-on-other-os</id>
+			<activation>
+				<os>
+					<family>!mac</family>
+				</os>
+			</activation>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.eclipse.tycho</groupId>
+						<artifactId>target-platform-configuration</artifactId>
+						<version>${tychoVersion}</version>
+						<configuration>
+							<dependency-resolution>
+								<optionalDependencies>ignore</optionalDependencies>
+							</dependency-resolution>
+						</configuration>
+					</plugin>
+					<plugin>
+						<groupId>org.eclipse.tycho</groupId>
+						<artifactId>tycho-compiler-plugin</artifactId>
+						<version>${tychoVersion}</version>
+						<configuration>
+							<optionalDependencies>ignore</optionalDependencies>
+						</configuration>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+		<profile>
+			<id>remove-other-platforms-from-query-on-mac</id>
+			<activation>
+				<os>
+					<family>mac</family>
+				</os>
+			</activation>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.eclipse.tycho</groupId>
+						<artifactId>target-platform-configuration</artifactId>
+						<version>${tychoVersion}</version>
+						<configuration>
+							<environments>
+								<environment>
+									<os>macosx</os>
+									<ws>cocoa</ws>
+									<arch>x86_64</arch>
+								</environment>
+								<environment>
+									<os>macosx</os>
+									<ws>cocoa</ws>
+									<arch>x86</arch>
+								</environment>
+							</environments>
+						</configuration>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+	</profiles>
+</project>
